+++ R/zzz.R
@@ -1,6 +1,4 @@
 #' @import rlang
-#' @import dplyr
-#' @import tidyr
 NULL
 
 # nocov start